chore: moving type definations
